chore: restore subnav (#773)
